requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docutils==0.20
idna==3.4
lockfile==0.12.2
luigi==3.3.0
numpy==1.24.3
pandas==2.0.1
praw==7.7.0
prawcore==2.3.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.30.0
six==1.16.0
tenacity==8.2.2
tornado==6.3.1
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.2
websocket-client==1.5.1
